fix(react-hot-loader): fix missing export
