+++ packages/react-hot-loader/src/patch.dev.js
@@ -2,3 +2,5 @@ import React from 'react'
 import reactHotLoader from './reactHotLoader'
 
 reactHotLoader.patch(React)
+
+export default reactHotLoader
